requirements.txt
aria2p>=0.9.0,<0.15.0
beautifulsoup4>=4.8.2,<4.8.10
google-api-python-client>=1.7.11,<1.7.20
google-auth-httplib2>=0.0.3,<0.1.0
google-auth-oauthlib>=0.4.1,<0.10.0
httplib2==0.15.0
oauth2client
psutil
Pyrogram
python-dotenv>=0.10
python-magic
requests
tenacity>=6.0.0
TgCrypto>=1.1.1,<1.1.10
git+git://github.com/sudoshell/youtube-dl@master#youtube_dl
